Change 'took' type from int to double in Groups model

+++ lib/models/api/v6/groups/groups.dart
@@ -7,7 +7,7 @@ part 'groups.g.dart';
 sealed class Groups with _$Groups {
   const factory Groups({
     required List<Group> groups,
-    required int took,
+    required double took,
     Processed? processed,
   }) = _Groups;
